+++ cli/azd/extensions/azure.ai.projects/pkg/provisioning/foundry_provisioning_provider_brownfield_acr_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package project
+package provisioning
 
 import (
 	"context"
@@ -9,7 +9,7 @@ import (
 	"strings"
 	"testing"
 
-	"azureaiagent/internal/synthesis"
+	"azure.ai.projects/pkg/provisioning/synthesis"
 
 	"github.com/azure/azure-dev/cli/azd/pkg/azdext"
 	"github.com/stretchr/testify/assert"